update geometry and continue tidy up of utils.h
