+++ cpp/CMakeLists.txt
@@ -58,6 +58,7 @@ include(GNUInstallDirs)
 install(FILES Contact.hpp contact_kernels.hpp utils.h coefficients.h geometric_quantities.h DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dolfinx_contact COMPONENT Development)
 
 target_sources(dolfinx_contact PRIVATE
+  ${CMAKE_CURRENT_SOURCE_DIR}/utils.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/coefficients.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/geometric_quantities.cpp
   )
