+++ python/demos/nitsche_unbiased_demo.py
@@ -15,6 +15,7 @@ from dolfinx_contact.meshing import (convert_mesh, create_box_mesh_2D,
                                      create_circle_circle_mesh,
                                      create_circle_plane_mesh,
                                      create_sphere_plane_mesh)
+from dolfinx_contact import update_geometry
 from dolfinx_contact.unbiased.nitsche_unbiased import nitsche_unbiased
 
 if __name__ == "__main__":
@@ -275,10 +276,7 @@ if __name__ == "__main__":
             xdmf.write_function(u1)
 
         # Perturb mesh with solution displacement
-        delta_x = u1.compute_point_values()
-        if delta_x.shape[1] < 3:
-            delta_x = np.hstack([delta_x, np.zeros((delta_x.shape[0], 3 - delta_x.shape[1]))])
-        mesh.geometry.x[:] += delta_x
+        update_geometry(u1._cpp_object, mesh)
 
         # Accumulate displacements
         u.x.array[:] += u1.x.array[:]
